vscode-extension: by default, the ActionScript TaskProvider uses a problem matcher that doesn't try to match any problems.
The reason for this is that the language server should have already found all problems, so it will create a ton of duplicates. Additionally, once you start making changes to the code to fix problems, I can't seem to figure out how to make the problems from the task go away.