asconfigc: can publish Animate FLAs (references #305)
Moved this task into asconfigc because we need to watch for certain log files to be written to know when publishing is complete. This allows us to use the Animate publish command as a preLaunchTask to (optionally) debug the SWF inside VSCode. Debugging inside Animate will be a separate command.