language-server: smart detection of underscore character (_) when generating getters/setters (closes #450)
if the variable name already starts with an underscore, the new name of the variable will be the same, and the underscore will be removed from the name of the generated getter/setter. If the variable name doesn't start with an underscore yet, an underscore will be added to the beginning of the variable name and the getter/setter will get the original name.