Add new status bar option to set preferred target in Royale projects (closes #155)
The language server previously always selected the first value of the `targets` compiler option. To switch between COMPILE::JS and COMPILE::SWF, users were required to modify asconfig.json to change the order of `targets`. However, when building libraries, the order of the targets is important, and this could break builds if the targets were in the wrong order. Now, there's a new status bar item you can use to select one of the options from the targets compiler option as your preferred target. This value is saved on a per-workspace basis, and is saved between sessions. If asconfig.json is modified, the preferred target may be reset to the default (JSRoyale, or another JS target, if available) if it no longer exists.

In non-Royale projects, this status bar item is not shown and changing the value has no effect.

Saved as a vscode.Memento instead of a setting.