vscode-extension: ActionScript Source Paths view and SDK/Royale status bar items will be displayed and remain open after the AS3 & MXML language server activates its first project for a workspace (closes #742) (references #351)
Before a project is created in a workspace (or when no workspace is currently open), these views will remain hidden unless a file with the extension .as or .mxml, or an asconfig.json or asconfig.*.json file, is open in a visible editor. They will be automatically hidden again after switching all visible editors to other file types.

This should strike a better balance. Inside a project where AS3 and MXML are the dominant languages, everything will always be visible, so there will be no confusion about why things are disappearing sometimes. In projects where other languages are dominant, this extension won't be annoying with things that probably aren't relevant.