language-server: improve performance of real-time problems checker
Use custom AS and MXML compilation units that override startBuildAsync() and get only the syntax tree and file scope requests. The others are skipped in that method, but they should still get triggered eventually.

What seemed to be happening is that the other requests triggered more requests in dependency compilation units, which was very expensive. The requests on the other compilation units can wait until a later point when they won't slow down real time problems.

This commit requires commit 8773198, or it won't help very much because if the get() calls on the requests aren't skipped, it will again force more requests on other dependency compilation units.